refactor: give stop-aware pipe reads an owner
